requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pandas>=1.2.0
Pillow>=8.0.0
tqdm>=4.50.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
wandb>=0.10.0 